Release 1.0

+++ app/src/main/java/com/courierx/courierx/Models/CreditLog.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @Data
 public class CreditLog {
     private String date;
-    private Float amount;
+    private Long amount;
     private String type;
 }
